+++ src/Plugin.php
@@ -13,11 +13,13 @@ use craft\events\RebuildConfigEvent;
 use craft\events\RegisterUrlRulesEvent;
 use craft\events\RegisterUserPermissionsEvent;
 use craft\helpers\UrlHelper;
-use craft\log\FileTarget;
+use craft\log\MonologTarget;
 use craft\services\ProjectConfig;
 use craft\services\UserPermissions;
 use craft\web\twig\variables\CraftVariable;
 use craft\web\UrlManager;
+use Monolog\Formatter\LineFormatter;
+use Psr\Log\LogLevel;
 use venveo\oauthclient\services\Apps as AppsService;
 use venveo\oauthclient\services\Credentials as CredentialsService;
 use venveo\oauthclient\services\Providers;
@@ -42,29 +44,14 @@ use yii\base\Event;
  */
 class Plugin extends BasePlugin
 {
-    const HANDLE = 'oauthclient';
+    public const HANDLE = 'oauthclient';
 
-    // Static Properties
-    // =========================================================================
-    public static $PROJECT_CONFIG_KEY = 'oauthClient';
+    public static string $PROJECT_CONFIG_KEY = 'oauthClient';
+
+    public string $schemaVersion = '2.1.3';
+    public bool $hasCpSettings = true;
 
 
-    /**
-     * @var Plugin
-     */
-    public static $plugin;
-
-    // Public Properties
-    // =========================================================================
-
-    /**
-     * @var string
-     */
-    public $schemaVersion = '2.1.3';
-    public $hasCpSettings = true;
-
-    // Public Methods
-    // =========================================================================
 
     /**
      * @inheritdoc
@@ -72,14 +59,12 @@ class Plugin extends BasePlugin
     public function init()
     {
         parent::init();
-        self::$plugin = $this;
 
         if (Craft::$app->request->getIsConsoleRequest()) {
             $this->controllerNamespace = 'venveo\oauthclient\console\controllers';
         }
 
         $this->_registerLogger();
-        $this->_setComponents();
         $this->_registerCpRoutes();
         $this->_registerVariables();
         $this->_registerProjectConfig();
@@ -94,28 +79,35 @@ class Plugin extends BasePlugin
      */
     private function _registerLogger()
     {
-        Craft::getLogger()->dispatcher->targets[] = new FileTarget([
-            'logFile' => Craft::getAlias('@storage/logs/oauthclient.log'),
-            'categories' => ['venveo\oauthclient\*'],
+        Craft::getLogger()->dispatcher->targets[] = new MonologTarget([
+            'name' => 'oauthclient',
+            'categories' => ['venveo\oauthcleint\*'],
+            'level' => LogLevel::INFO,
+            'logContext' => false,
+            'allowLineBreaks' => false,
+            'formatter' => new LineFormatter(
+                format: "[%datetime%] %message%\n",
+                dateFormat: 'Y-m-d H:i:s',
+            ),
         ]);
     }
 
-    // Private Methods
-    // =========================================================================
 
     /**
-     * Set our service components
+     * @inheritdoc
      */
-    private function _setComponents()
+    public static function config(): array
     {
-        $this->setComponents([
-            'apps' => AppsService::class,
-            'providers' => ProvidersService::class,
-            'tokens' => TokensService::class,
-            'credentials' => CredentialsService::class,
-        ]);
-
+        return [
+            'components' => [
+                'apps' => ['class' => AppsService::class],
+                'providers' => ['class' => ProvidersService::class],
+                'tokens' => ['class' => TokensService::class],
+                'credentials' => ['class' => CredentialsService::class]
+            ],
+        ];
     }
+
 
     /**
      * Adds the event handler for registering CP routes
@@ -130,11 +122,6 @@ class Plugin extends BasePlugin
                 'oauthclient/apps/<handle:{handle}>' => 'oauthclient/apps/edit',
                 'oauthclient/apps/delete' => 'oauthclient/apps/delete',
 
-                // TODO: Remove these in next version in favor of `oauth` route
-                'oauthclient/authorize/refresh/<id:\d+>' => 'oauthclient/authorize/refresh',
-                'oauthclient/authorize/<handle:{handle}>' => 'oauthclient/authorize/authorize-app',
-                // These are duplicates of potentially non-admin-facing actions. Craft automatically checks routes for
-                // the plugin handle, so we needed new routes without the handle.
                 'oauth/authorize/refresh/<id:\d+>' => 'oauthclient/authorize/refresh',
                 'oauth/authorize/<handle:{handle}>' => 'oauthclient/authorize/authorize-app',
             ]);
@@ -197,16 +184,19 @@ class Plugin extends BasePlugin
     private function _registerPermissions()
     {
         Event::on(UserPermissions::class, UserPermissions::EVENT_REGISTER_PERMISSIONS, function (RegisterUserPermissionsEvent $event) {
-            $apps = Plugin::$plugin->apps->getAllApps();
+            $apps = $this->apps->getAllApps();
             $loginPermissions = [];
             foreach ($apps as $app) {
                 $suffix = ':' . $app->uid;
                 $loginPermissions['oauthclient-login' . $suffix] = ['label' => self::t('Login to “{name}” ({handle}) app', ['name' => $app->name, 'handle' => $app->handle])];
             }
-            $event->permissions[self::t('OAuth Client')] = [
-                'oauthclient-login' => [
-                    'label' => self::t('Login to Apps'), 'nested' => $loginPermissions
-                ]
+            $event->permissions[] = [
+                'heading' => self::t('OAuth Client'),
+                'permissions' => [
+                    'oauthclient-login' => [
+                        'label' => self::t('Login to Apps'), 'nested' => $loginPermissions
+                    ]
+                ],
             ];
         });
     }
@@ -222,7 +212,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public function getSettingsResponse()
+    public function getSettingsResponse(): mixed
     {
         return Craft::$app->getResponse()->redirect(UrlHelper::cpUrl('oauthclient/apps'));
     }