+++ src/variables/OAuthVariable.php
@@ -20,6 +20,6 @@ class OAuthVariable
      */
     public function getAppByHandle($handle)
     {
-        return Plugin::$plugin->apps->getAppByHandle($handle);
+        return Plugin::getInstance()->apps->getAppByHandle($handle);
     }
 }